refactor(credit): Phase 1c — assemble deep_review/ package from staging
- deep_review_pkg/ → deep_review/ (Git tracks renames)
- Created __init__.py with PEP 562 lazy imports (matches edgar/memo pattern)
- git rm'd 6 flat deep_review*.py files
- Updated all internal imports: deep_review_helpers → deep_review.helpers, etc.
- 11 modules: __init__, models, helpers, corpus, prompts, policy, decision,
  confidence, persist, portfolio, service
