+++ backend/vertical_engines/credit/deep_review/__init__.py
@@ -1,0 +1,87 @@
+"""Deep Review V4 — IC-Grade Investment Memorandum Pipeline.
+
+Three-tier DAG: models → domain modules → persist/portfolio → service.
+This is distinct from the standard two-tier edgar pattern because deep_review
+is the only engine that persists other engines' results.
+
+Public API:
+    run_deal_deep_review_v4()           — sync single-deal deep review
+    async_run_deal_deep_review_v4()     — async single-deal deep review
+    run_all_deals_deep_review_v4()      — sync batch deep review
+    async_run_all_deals_deep_review_v4() — async batch deep review
+    run_portfolio_review()              — single investment periodic review
+    run_all_portfolio_reviews()         — batch portfolio reviews
+    get_current_im_draft()              — query current IM draft
+    compute_underwriting_confidence()   — underwriting reliability score
+    apply_tone_normalizer_adjustment()  — post-tone confidence adjustment
+
+Error contract: never-raises (orchestration engine). Functions return result
+dicts with status/warnings on failure. exc_info=True in structlog.
+"""
+from __future__ import annotations
+
+from typing import TYPE_CHECKING, Any
+
+# Eagerly import leaf-node symbols only (no transitive deps)
+from vertical_engines.credit.deep_review.confidence import (
+    apply_tone_normalizer_adjustment,
+    compute_underwriting_confidence,
+)
+
+if TYPE_CHECKING:
+    from vertical_engines.credit.deep_review.portfolio import (
+        get_current_im_draft as get_current_im_draft,
+        run_all_portfolio_reviews as run_all_portfolio_reviews,
+        run_portfolio_review as run_portfolio_review,
+    )
+    from vertical_engines.credit.deep_review.service import (
+        async_run_all_deals_deep_review_v4 as async_run_all_deals_deep_review_v4,
+        async_run_deal_deep_review_v4 as async_run_deal_deep_review_v4,
+        run_all_deals_deep_review_v4 as run_all_deals_deep_review_v4,
+        run_deal_deep_review_v4 as run_deal_deep_review_v4,
+    )
+
+__all__ = [
+    "run_deal_deep_review_v4",
+    "async_run_deal_deep_review_v4",
+    "run_all_deals_deep_review_v4",
+    "async_run_all_deals_deep_review_v4",
+    "run_portfolio_review",
+    "run_all_portfolio_reviews",
+    "get_current_im_draft",
+    "compute_underwriting_confidence",
+    "apply_tone_normalizer_adjustment",
+]
+
+
+def __getattr__(name: str) -> Any:
+    if name in (
+        "run_deal_deep_review_v4",
+        "async_run_deal_deep_review_v4",
+        "run_all_deals_deep_review_v4",
+        "async_run_all_deals_deep_review_v4",
+    ):
+        from vertical_engines.credit.deep_review.service import (
+            async_run_all_deals_deep_review_v4,
+            async_run_deal_deep_review_v4,
+            run_all_deals_deep_review_v4,
+            run_deal_deep_review_v4,
+        )
+        return {
+            "run_deal_deep_review_v4": run_deal_deep_review_v4,
+            "async_run_deal_deep_review_v4": async_run_deal_deep_review_v4,
+            "run_all_deals_deep_review_v4": run_all_deals_deep_review_v4,
+            "async_run_all_deals_deep_review_v4": async_run_all_deals_deep_review_v4,
+        }[name]
+    if name in ("run_portfolio_review", "run_all_portfolio_reviews", "get_current_im_draft"):
+        from vertical_engines.credit.deep_review.portfolio import (
+            get_current_im_draft,
+            run_all_portfolio_reviews,
+            run_portfolio_review,
+        )
+        return {
+            "run_portfolio_review": run_portfolio_review,
+            "run_all_portfolio_reviews": run_all_portfolio_reviews,
+            "get_current_im_draft": get_current_im_draft,
+        }[name]
+    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
